+++ src/test/java/de/codescape/jira/plugins/scrumpoker/rest/entities/CardEntityTest.java
@@ -6,15 +6,15 @@ import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertThat;
 
-public class CardRepresentationTest {
+public class CardEntityTest {
 
     private static final String CARD_VALUE = "5";
 
     @Test
     public void cardReturnsValuesItIsInitializedWith() {
-        CardRepresentation cardRepresentation = new CardRepresentation(CARD_VALUE, true);
-        assertThat(cardRepresentation.getValue(), is(equalTo(CARD_VALUE)));
-        assertThat(cardRepresentation.isSelected(), is(equalTo(true)));
+        CardEntity cardEntity = new CardEntity(CARD_VALUE, true);
+        assertThat(cardEntity.getValue(), is(equalTo(CARD_VALUE)));
+        assertThat(cardEntity.isSelected(), is(equalTo(true)));
     }
 
 }