+++ src/test/java/de/codescape/jira/plugins/scrumpoker/rest/entities/VoteEntityTest.java
@@ -6,17 +6,17 @@ import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertThat;
 
-public class VoteRepresentationTest {
+public class VoteEntityTest {
 
     private static final String USER = "Some User";
     private static final String VOTE = "?";
 
     @Test
     public void voteReturnsValuesItIsInitializedWith() {
-        VoteRepresentation voteRepresentation = new VoteRepresentation(USER, VOTE, true);
-        assertThat(voteRepresentation.getUser(), is(equalTo(USER)));
-        assertThat(voteRepresentation.getVote(), is(equalTo(VOTE)));
-        assertThat(voteRepresentation.isNeedToTalk(), is(equalTo(true)));
+        VoteEntity voteEntity = new VoteEntity(USER, VOTE, true);
+        assertThat(voteEntity.getUser(), is(equalTo(USER)));
+        assertThat(voteEntity.getVote(), is(equalTo(VOTE)));
+        assertThat(voteEntity.isNeedToTalk(), is(equalTo(true)));
     }
 
 }